refactor: mv src/gui/src/ScreenSettingsDialog => src/gui/src/dialogs/ScreenSettingsDialog
