+++ src/gui/src/dialogs/ScreenSettingsDialog.cpp
