+++ src/gui/src/dialogs/ScreenSettingsDialog.h
