+++ src/gui/src/dialogs/ScreenSettingsDialog.ui
